feat(git): add denops argument to getGitRoot
This commit adds a `denops` argument to the `getGitRoot` function. This allows the function to be used in a Denops context, where it can access the current working directory.

The previous implementation of `getGitRoot` used the `Deno.cwd()` function to get the current working directory. However, this function is not available in a Denops context. Therefore, the `getGitRoot` function has been updated to use the `fn.getcwd(denops)` function instead.

This change allows the `gitCommit` function to be used in a Denops context. The `gitCommit` function uses the `getGitRoot` function to get the current working directory. This information is used to generate the commit message.